e2e test cleanup
