+++ packages/app/tests/e2e/features/search/search.spec.ts
@@ -117,128 +117,6 @@ test.describe('Search', { tag: '@search' }, () => {
       });
     });
 
-    test('should support multiline SQL WHERE clauses', async ({ page }) => {
-      const searchForm = page.locator('[data-testid="search-form"]');
-      const sqlToggle = searchForm.locator('text=SQL').first();
-      const whereContainer = searchForm.locator('div:has(p.mantine-Text-root:has-text("WHERE"))');
-      const whereLabel = whereContainer.locator('p.mantine-Text-root:has-text("WHERE")');
-      const whereEditor = whereContainer.locator('.cm-editor');
-      const searchResults = page.locator('[data-testid="search-results-table"]');
-
-      await test.step('Switch to SQL mode', async () => {
-        await sqlToggle.click();
-        await expect(whereLabel).toBeVisible();
-      });
-
-      await test.step('Test multiline input with Shift+Enter', async () => {
-        await whereEditor.click();
-        
-        await whereEditor.type('timestamp >= now() - interval 1 hour');
-        await page.keyboard.press('Shift+Enter');
-        await whereEditor.type('AND level = "error"');
-        await page.keyboard.press('Shift+Enter'); 
-        await whereEditor.type('AND service_name = "api"');
-        
-        const editorContent = await whereEditor.textContent();
-        expect(editorContent).toContain('timestamp >= now() - interval 1 hour');
-        expect(editorContent).toContain('AND level = "error"');
-        expect(editorContent).toContain('AND service_name = "api"');
-      });
-
-      await test.step('Test editor height expansion', async () => {
-        const initialBox = await whereEditor.boundingBox();
-        const initialHeight = initialBox?.height || 0;
-        
-        await whereEditor.press('Shift+Enter');
-        await whereEditor.type('AND response_time > 1000');
-        await whereEditor.press('Shift+Enter');
-        await whereEditor.type('AND user_id IS NOT NULL');
-        
-        const expandedBox = await whereEditor.boundingBox();
-        const expandedHeight = expandedBox?.height || 0;
-        expect(expandedHeight).toBeGreaterThan(initialHeight);
-      });
-
-      await test.step('Test max height with scroll overflow', async () => {
-        for (let i = 0; i < 10; i++) {
-          await whereEditor.press('Shift+Enter');
-          await whereEditor.type(`AND field_${i} = "value_${i}"`);
-        }
-        
-        const editorBox = await whereEditor.boundingBox();
-        const maxHeight = 150;
-        expect(editorBox?.height).toBeLessThanOrEqual(maxHeight + 10);
-        
-        const scroller = whereEditor.locator('.cm-scroller');
-        await expect(scroller).toHaveCSS('overflow-y', 'auto');
-      });
-
-      await test.step('Test Enter submits query', async () => {
-        await whereEditor.click();
-        
-        await page.keyboard.press('Control+a');
-        await whereEditor.type('level = "info"');
-        
-        await page.keyboard.press('Enter');
-        
-        await page.waitForTimeout(1000);
-        await expect(searchResults).toBeVisible({ timeout: 5000 });
-      });
-    });
-
-    test('should support multiline Lucene search input', async ({ page }) => {
-      const searchForm = page.locator('[data-testid="search-form"]');
-      const luceneToggle = searchForm.locator('text=Lucene').first();
-      const searchInput = page.locator('[data-testid="search-input"]');
-      const searchResults = page.locator('[data-testid="search-results-table"]');
-
-      await test.step('Ensure Lucene mode is active', async () => {
-        await luceneToggle.click();
-        await expect(searchInput).toBeVisible();
-      });
-
-      await test.step('Test multiline input with auto-expansion', async () => {
-        await searchInput.click();
-        
-        await searchInput.type('level:error');
-        await page.keyboard.press('Shift+Enter');
-        await searchInput.type('service_name:api');
-        await page.keyboard.press('Shift+Enter');
-        await searchInput.type('timestamp:[now-1h TO now]');
-        
-        const inputValue = await searchInput.inputValue();
-        expect(inputValue).toContain('level:error');
-        expect(inputValue).toContain('service_name:api');
-        expect(inputValue).toContain('timestamp:[now-1h TO now]');
-      });
-
-      await test.step('Test textarea auto-expansion', async () => {
-        const initialBox = await searchInput.boundingBox();
-        const initialHeight = initialBox?.height || 0;
-        
-        await page.keyboard.press('Shift+Enter');
-        await searchInput.type('response_time:>1000');
-        await page.keyboard.press('Shift+Enter');
-        await searchInput.type('user_id:*');
-        
-        const expandedBox = await searchInput.boundingBox();
-        const expandedHeight = expandedBox?.height || 0;
-        expect(expandedHeight).toBeGreaterThan(initialHeight);
-      });
-
-      await test.step('Test Enter submits search', async () => {
-        await searchInput.click();
-        
-        await page.keyboard.press('Control+a');
-        await searchInput.type('level:info');
-        
-        await page.keyboard.press('Enter');
-        
-        await page.waitForTimeout(1000);
-        await expect(searchResults).toBeVisible({ timeout: 5000 });
-      });
-    });
-
     test('Comprehensive Search Workflow - Search, View Results, Navigate Side Panel', async ({
       page,
     }) => {
